requirements.txt
boto3==1.42.38
botocore==1.42.38
certifi==2026.1.4
charset-normalizer==3.4.4
Faker==40.1.2
idna==3.11
jmespath==1.1.0
numpy==2.4.1
pandas==3.0.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
urllib3==2.6.3
